requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
attrs==23.2.0
azure-core==1.34.0
azure-identity==1.15.0
azure-servicebus==7.14.2
bidict==0.23.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
dnspython==2.6.1
eventlet==0.36.1
flower==2.0.1
frozenlist==1.4.1
googleapis-common-protos==1.59.1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
humanize==4.9.0
ibm-cloud-sdk-core==3.20.0
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-watson==8.0.0
ibm_watson_machine_learning==1.0.356
idna==3.6
importlib-metadata==7.0.0
isodate==0.7.2
jmespath==1.0.1
joblib==1.4.2
kombu==5.3.5
lomond==0.3.3
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-celery==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
packaging==24.0
pandas==2.1.4
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
redis==5.0.1
regex==2024.5.15
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
simple-websocket==1.0.0
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.5.0
thrift==0.20.0
tornado==6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.7.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
pika==1.3.2
